requirements.txt
datasets>=1.11.0
flatten-dict
hydra-core>=1.0.6
isort==5.9.3
librosa
pycocotools
pylint==2.11.0
pytorch-lightning==1.5.0
scipy
toml==0.10.2
torch==1.9.0
torchaudio==0.9.0
torchmetrics>=0.3.0
torchvision==0.10.0
tqdm
transformers
wandb
wget
yapf==0.31.0
einops
dotmap
pandas
nlpaug
matplotlib
torch_dct

pytorch-lightning-bolts==0.3.2